+++ core_client/base/models/v3/report_process_list.py
@@ -1,16 +1,6 @@
-from pydantic import BaseModel, RootModel
+from pydantic import RootModel
 
 from . import ReportProcess
 
-
-class ReportProcessList(BaseModel):
-    """
-    [{
-        "created_at": 0,
-        "exit_state": "string",
-        "id": "string",
-        "reference": "string"
-    }]
-    """
-
-    RootModel: ReportProcess
+# JSON array of ReportProcess.
+ReportProcessList = RootModel[list[ReportProcess]]